# src/middleware/openapi.yml
openapi: 3.0.1
info:
  title: Companies API
  description: This is the Companies API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.3.0
paths:
  /:
    get:
      summary: Returns available resources
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Root'
  /companies:
    get:
      summary: Returns a list of all companies (subject to LIMIT value)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Companies'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Creates a new company
      requestBody:
        description: Company to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCompany'
      responses:
        '201':
          description: Successful created new company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
  /companies/{companyId}:
    get:
      summary: Gets data about a company
      responses:
        '200':
          description: Gets data about a {companyId} company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  responses:
    BadRequest:
      description: Unsuccessful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Unsuccessful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'  
  schemas:
    Root:
      type: object
    Companies:
      type: array
      items:
        type: object
    NewCompany:
      type: object
      description: New company object
      properties:
        company_id:
          type: string
        name:
          type: string
        summary:
          type: string
        careers:
          type: string
        industry:
          type: string
        founded:
          type: integer
        github:
          type: string
        blog:
          type: string
      required:
        - company_id
        - name
        - summary
        - careers
        - industry
        - founded
        - github
        - blog
    Company:
      type: object
      properties:
        company_id:
          type: string
        name:
          type: string
        summary:
          type: string
        careers:
          type: string
        industry:
          type: string
        founded:
          type: integer
        github:
          type: string
        blog:
          type: string
    Error:
      type: object
      properties:
        title:
          description: A short, human-readable title for the general error type; the title should not change for given types
          type: string
        status:
          description: HTTP status code
          type: integer
        instance:
          description: A URI reference that identifies the specific occurrence of the problem
          type: string
        detail:
          description: A human-readable description of the specific error
          type: string